+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.6.0 (TBD)
 
+- [BREAKING] Changed `TransactionExecutor` and `TransactionHost` to use trait objects (#897).
 - Implemented kernel procedure to set transaction expiration block delta (#897).
 - Created a proving service that receives `TransactionWitness` and returns the proof using gRPC (#881).
 - Made note scripts public (#880).
@@ -23,6 +24,7 @@
 - Introduced `AccountDelta` maximum size limit of 32 KiB (#889).
 - [BREAKING] Moved `MAX_NUM_FOREIGN_ACCOUNTS` into `miden-objects` (#904).
 - Implemented `storage_size`, updated storage bounds (#886).
+- [BREAKING] Auto-generate `KERNEL_ERRORS` list from the transaction kernel's MASM files and rework error constant names (#906).
 
 ## 0.5.1 (2024-08-28) - `miden-objects` crate only